+++ include/exec/sequence_senders.hpp
@@ -145,268 +145,6 @@ namespace exec {
          { sequence_connect((_Sender&&) __sndr, (_Receiver&&) __rcvr) } -> stdexec::operation_state;
        };
 
-  namespace __repeat {
-    using namespace stdexec;
-
-    // Takes a sender and creates a sequence sender by repeating the sender as item to the set_next
-    // operation.
-    template <class _SourceSenderId, class _ReceiverId>
-    struct __operation {
-      using _SourceSender = stdexec::__t<_SourceSenderId>;
-      using _Receiver = stdexec::__t<_ReceiverId>;
-      using __next_sender_t = __call_result_t<set_next_t, _Receiver&, _SourceSender>;
-      using __next_on_scheduler_sender =
-        __call_result_t<on_t, exec::trampoline_scheduler, __next_sender_t&&>;
-
-      struct __t;
-
-      struct __receiver {
-        __t* __op_;
-
-        friend void tag_invoke(set_value_t, __receiver&& __self) noexcept {
-          __self.__op_->repeat();
-        }
-
-        template <__decays_to<__receiver> _Self, class _Error>
-          requires __callable<set_error_t, _Receiver&&, _Error&&>
-        friend void tag_invoke(set_error_t, _Self&& __self, _Error e) noexcept {
-          stdexec::set_error((_Receiver&&) __self.__op_->__rcvr_, (_Error&&) e);
-        }
-
-        friend void tag_invoke(set_stopped_t, __receiver&& __self) noexcept {
-          auto token = stdexec::get_stop_token(stdexec::get_env(__self.__op_->__rcvr_));
-          if (token.stop_requested()) {
-            stdexec::set_stopped((_Receiver&&) __self.__op_->__rcvr_);
-          } else {
-            stdexec::set_value((_Receiver&&) __self.__op_->__rcvr_);
-          }
-        }
-
-        friend env_of_t<_Receiver> tag_invoke(get_env_t, const __receiver& __self) noexcept(
-          __nothrow_callable<get_env_t, const _Receiver&>) {
-          return stdexec::get_env(__self.__op_->__rcvr_);
-        }
-      };
-
-      struct __t {
-        using __id = __operation;
-        [[no_unique_address]] _Receiver __rcvr_;
-        [[no_unique_address]] _SourceSender __source_;
-        exec::trampoline_scheduler __trampoline_;
-        std::optional<connect_result_t<__next_on_scheduler_sender, __receiver>> __next_op_;
-
-        void repeat() noexcept {
-          auto __token = stdexec::get_stop_token(stdexec::get_env(__rcvr_));
-          if (__token.stop_requested()) {
-            stdexec::set_stopped((_Receiver&&) __rcvr_);
-            return;
-          }
-          try {
-            auto& __next = __next_op_.emplace(__conv{[&] {
-              return connect(
-                stdexec::on(__trampoline_, set_next(__rcvr_, _SourceSender{__source_})),
-                __receiver{this});
-            }});
-            stdexec::start(__next);
-          } catch (...) {
-            stdexec::set_error((_Receiver&&) __rcvr_, std::current_exception());
-          }
-        }
-
-        friend void tag_invoke(start_t, __t& __self) noexcept {
-          __self.repeat();
-        }
-
-        template <__decays_to<_SourceSender> _Sndr, __decays_to<_Receiver> _Rcvr>
-        explicit __t(_Sndr&& __source, _Rcvr&& __rcvr)
-          : __rcvr_{(_Rcvr&&) __rcvr}
-          , __source_{(_Sndr&&) __source} {
-        }
-      };
-    };
-
-    template <class _SourceSender, class _Receiver>
-    using __operation_t = __t<__operation<__id<_SourceSender>, __id<__decay_t<_Receiver>>>>;
-
-    template <class _Source, class _Env>
-    using __compl_sigs = make_completion_signatures<
-      _Source,
-      _Env,
-      completion_signatures<set_error_t(std::exception_ptr), set_stopped_t()>>;
-
-    template <class _SourceId>
-    struct __sender {
-      using _Source = stdexec::__t<__decay_t<_SourceId>>;
-
-      template <class _Rcvr>
-      using __next_sender = __call_result_t<set_next_t, __decay_t<_Rcvr>&, _Source>;
-
-      template <class _Rcvr>
-      using __next_on_scheduler_sender =
-        __call_result_t<on_t, exec::trampoline_scheduler, __next_sender<_Rcvr>&&>;
-
-      template <class _Rcvr>
-      using __recveiver = typename __operation<__id<_Source>, __id<__decay_t<_Rcvr>>>::__receiver;
-
-      class __t {
-        [[no_unique_address]] _Source __source_;
-
-        template <__decays_to<__t> _Self, class _Receiver>
-          requires sequence_receiver_of<_Receiver, __compl_sigs<_Source&&, env_of_t<_Receiver>>>
-                && __valid<__next_on_scheduler_sender, _Receiver>
-                && sender_to<__next_on_scheduler_sender<_Receiver>, __recveiver<_Receiver>>
-        friend auto tag_invoke(sequence_connect_t, _Self&& __self, _Receiver&& __rcvr)
-          -> __operation_t<_Source, _Receiver> {
-          return __operation_t<_Source, _Receiver>{
-            ((_Self&&) __self).__source_, (_Receiver&&) __rcvr};
-        }
-
-        template <__decays_to<__t> _Self, class _Env>
-        friend auto tag_invoke(get_completion_signatures_t, _Self&&, const _Env&)
-          -> __compl_sigs<_Source&&, _Env>;
-
-       public:
-        using __id = __sender;
-
-        template <__decays_to<_Source> _Sndr>
-        explicit __t(_Sndr&& __source)
-          : __source_((_Sndr&&) __source) {
-        }
-      };
-    };
-
-    struct __repeat_t {
-      template <sender Sender>
-      auto operator()(Sender&& source) const {
-        return __t<__sender<__id<__decay_t<Sender>>>>{static_cast<Sender&&>(source)};
-      }
-    };
-  } // namespace repeat_
-
-  using __repeat::__repeat_t;
-  inline constexpr __repeat_t repeat;
-
-  namespace __transform_each {
-    using namespace stdexec;
-
-    template <class _Receiver, class _Fun>
-    struct __operation_base {
-      [[no_unique_address]] _Receiver __rcvr_;
-      [[no_unique_address]] _Fun __fun_;
-    };
-
-    template <class _ReceiverId, class _Fun>
-    struct __receiver {
-      using _Receiver = stdexec::__t<_ReceiverId>;
-
-      struct __t {
-        __operation_base<_Receiver, _Fun>* __op_;
-
-        template <class _Item>
-          requires __callable<set_next_t, _Receiver&, __call_result_t<_Fun, _Item&&>>
-        friend auto tag_invoke(set_next_t, __t& __self, _Item&& __item) noexcept {
-          return set_next(
-            __self.__op_->__rcvr_, __self.__op_->__fun_(static_cast<_Item&&>(__item)));
-        }
-
-        template <__completion_tag _Tag, class... _Args>
-          requires __callable<_Tag, _Receiver&&, _Args&&...>
-        friend void tag_invoke(_Tag __complete, __t&& __self, _Args&&... __args) noexcept {
-          __complete(
-            static_cast<_Receiver&&>(__self.__op_->__rcvr_), static_cast<_Args&&>(__args)...);
-        }
-
-        friend env_of_t<_Receiver> tag_invoke(get_env_t, const __t& __self) noexcept {
-          return get_env(__self.__op_->__rcvr_);
-        }
-      };
-    };
-    template <class _Receiver, class _Fun>
-    using __receiver_t = __t<__receiver<__id<__decay_t<_Receiver>>, _Fun>>;
-
-    template <class _SenderId, class _ReceiverId, class _Fun>
-    struct __operation {
-      using _Sender = stdexec::__t<_SenderId>;
-      using _Receiver = stdexec::__t<_ReceiverId>;
-
-      struct __t : __operation_base<_Receiver, _Fun> {
-        sequence_connect_result_t<_Sender, __receiver_t<_Receiver, _Fun>> __op_;
-
-        template <class _Sndr, class _Rcvr>
-        __t(_Sndr&& __sndr, _Rcvr&& __rcvr, _Fun __fun)
-          : __operation_base<
-            _Receiver,
-            _Fun>{static_cast<_Rcvr&&>(__rcvr), static_cast<_Fun&&>(__fun)}
-          , __op_(
-              sequence_connect(static_cast<_Sndr&&>(__sndr), __receiver_t<_Receiver, _Fun>{this})) {
-        }
-
-       private:
-        friend void tag_invoke(start_t, __t& __self) noexcept {
-          start(__self.__op_);
-        }
-      };
-    };
-    template <class _Sender, class _Receiver, class _Fun>
-    using __operation_t = __t<__operation<__id<_Sender>, __id<__decay_t<_Receiver>>, _Fun>>;
-
-    template <class... _Errors>
-    using __error_sigs = completion_signatures<set_error_t(_Errors)...>;
-
-    template <class _Sender, class _Env, class _Fun>
-    using __completion_sigs = completion_signatures_of_t<__call_result_t<_Fun, _Sender>, _Env>;
-
-    template <class _Sender, class _Fun>
-    struct __sender {
-      struct __t {
-        [[no_unique_address]] _Sender __sndr_;
-        [[no_unique_address]] _Fun __fun_;
-
-        template <__decays_to<_Sender> Sndr>
-        __t(Sndr&& __sndr, _Fun __fun)
-          : __sndr_{static_cast<Sndr&&>(__sndr)}
-          , __fun_{static_cast<_Fun&&>(__fun)} {
-        }
-
-        template <__decays_to<__t> Self, class Rcvr>
-          requires sequence_receiver_of<
-                     Rcvr,
-                     __completion_sigs<__copy_cvref_t<Self, _Sender>, env_of_t<Rcvr>, _Fun>>
-                && sequence_sender_to<
-                     __copy_cvref_t<Self, _Sender>,
-                     __receiver_t<__decay_t<Rcvr>, _Fun>>
-        friend __operation_t<__copy_cvref_t<Self, _Sender>, Rcvr, _Fun>
-          tag_invoke(sequence_connect_t, Self&& __self, Rcvr&& __rcvr) {
-          return __operation_t<__copy_cvref_t<Self, _Sender>, Rcvr, _Fun>(
-            static_cast<Self&&>(__self).__sndr_,
-            static_cast<Rcvr&&>(__rcvr),
-            static_cast<Self&&>(__self).__fun_);
-        }
-
-        template <__decays_to<__t> _Self, class _Env>
-        friend auto tag_invoke(get_completion_signatures_t, _Self&&, const _Env&)
-          -> __completion_sigs<__copy_cvref_t<_Self, _Sender>, _Env, _Fun>;
-      };
-    };
-
-    struct transform_each_t {
-      template <stdexec::sender _Sender, class _Fun>
-      auto operator()(_Sender&& sndr, _Fun __fun) const {
-        return __t<__sender<__decay_t<_Sender>, _Fun>>{
-          static_cast<_Sender&&>(sndr), static_cast<_Fun&&>(__fun)};
-      }
-
-      template <class _Fun>
-      constexpr auto operator()(_Fun __fun) const noexcept
-        -> __binder_back<transform_each_t, _Fun> {
-        return {{}, {}, {static_cast<_Fun&&>(__fun)}};
-      }
-    };
-  } // namespace __transform_each
-
-  using __transform_each::transform_each_t;
-  inline constexpr transform_each_t transform_each{};
-
   namespace __let_xxx_each {
     using namespace stdexec;
 
@@ -494,6 +232,8 @@ namespace exec {
     template <class _Sender, class _Fun, auto _Let>
     struct __sender {
       struct __t {
+        using __id = __sender;
+
         [[no_unique_address]] _Sender __sndr_;
         [[no_unique_address]] _Fun __fun_;
 
@@ -694,90 +434,5 @@ namespace exec {
   } // namespace __enumerate_each
 
   inline constexpr __enumerate_each::enumerate_each_t enumerate_each{};
-
-  namespace __ignore_all {
-    using namespace stdexec;
-
-    template <class _ReceiverId>
-    struct __receiver {
-      using _Receiver = stdexec::__t<_ReceiverId>;
-
-      struct __t {
-        using __id = __receiver;
-        [[no_unique_address]] _Receiver __rcvr_;
-
-        template <class _Item>
-        friend _Item&& tag_invoke(set_next_t, __t&, _Item&& __item) noexcept {
-          return static_cast<_Item&&>(__item);
-        }
-
-        template <class _Error>
-        friend void tag_invoke(set_error_t, __t&& __self, _Error&& __error) noexcept {
-          set_error(static_cast<_Receiver&&>(__self.__rcvr_), static_cast<_Error&&>(__error));
-        }
-
-        template <__one_of<set_stopped_t, set_value_t> _Tag>
-        friend void tag_invoke(_Tag complete, __t&& __self) noexcept {
-          complete(static_cast<_Receiver&&>(__self.__rcvr_));
-        }
-
-        friend env_of_t<_Receiver> tag_invoke(get_env_t, const __t& __self) noexcept {
-          return get_env(__self.__rcvr_);
-        }
-      };
-    };
-
-    template <class _Rcvr>
-    using __receiver_t = __t<__receiver<__id<__decay_t<_Rcvr>>>>;
-
-    template <class... _Args>
-    using __drop_value_args = completion_signatures<set_value_t()>;
-
-    template <class _Sender, class _Env>
-    using __completion_sigs = make_completion_signatures<
-      _Sender,
-      _Env,
-      completion_signatures<set_value_t()>,
-      __drop_value_args>;
-
-    template <class _SenderId>
-    struct __sender {
-      using _Sender = stdexec::__t<__decay_t<_SenderId>>;
-
-      struct __t {
-        using __id = __sender;
-        [[no_unique_address]] _Sender __sndr_;
-
-        template <__decays_to<__t> _Self, class _Receiver>
-          requires receiver_of<
-                     _Receiver,
-                     __completion_sigs<__copy_cvref_t<_Self, _Sender>, env_of_t<_Receiver>>>
-                && sequence_sender_to<__copy_cvref_t<_Self, _Sender>, __receiver_t<_Receiver>>
-        friend sequence_connect_result_t< __copy_cvref_t<_Self, _Sender>, __receiver_t<_Receiver>>
-          tag_invoke(connect_t, _Self&& __self, _Receiver&& __rcvr) {
-          return sequence_connect(
-            ((_Self&&) __self).__sndr_, __receiver_t<_Receiver>{static_cast<_Receiver&&>(__rcvr)});
-        }
-
-        template <__decays_to<__t> _Self, class _Env>
-        friend auto tag_invoke(get_completion_signatures_t, _Self&&, const _Env&)
-          -> __completion_sigs<__copy_cvref_t<_Self, _Sender>, _Env>;
-      };
-    };
-
-    struct ignore_all_t {
-      template <class _Sender>
-      constexpr auto operator()(_Sender&& __sndr) const {
-        return __t<__sender<__id<__decay_t<_Sender>>>>{static_cast<_Sender&&>(__sndr)};
-      }
-
-      constexpr auto operator()() const noexcept -> __binder_back<ignore_all_t> {
-        return {};
-      }
-    };
-  } // namespace __ignore_all
-
-  using __ignore_all::ignore_all_t;
-  inline constexpr ignore_all_t ignore_all;
 
 }
